+++ python/ray/rllib/optimizers/__init__.py
@@ -1,14 +1,15 @@
-from ray.rllib.optimizers.apex_optimizer import ApexOptimizer
-from ray.rllib.optimizers.async_optimizer import AsyncOptimizer
-from ray.rllib.optimizers.local_sync import LocalSyncOptimizer
-from ray.rllib.optimizers.local_sync_replay import LocalSyncReplayOptimizer
-from ray.rllib.optimizers.multi_gpu import LocalMultiGPUOptimizer
+from ray.rllib.optimizers.async_samples_optimizer import AsyncSamplesOptimizer
+from ray.rllib.optimizers.async_gradients_optimizer import \
+    AsyncGradientsOptimizer
+from ray.rllib.optimizers.sync_samples_optimizer import SyncSamplesOptimizer
+from ray.rllib.optimizers.sync_replay_optimizer import SyncReplayOptimizer
+from ray.rllib.optimizers.multi_gpu_optimizer import LocalMultiGPUOptimizer
 from ray.rllib.optimizers.sample_batch import SampleBatch, MultiAgentBatch
 from ray.rllib.optimizers.policy_evaluator import PolicyEvaluator, \
     TFMultiGPUSupport
 
 
 __all__ = [
-    "ApexOptimizer", "AsyncOptimizer", "LocalSyncOptimizer",
-    "LocalSyncReplayOptimizer", "LocalMultiGPUOptimizer", "SampleBatch",
+    "AsyncSamplesOptimizer", "AsyncGradientsOptimizer", "SyncSamplesOptimizer",
+    "SyncReplayOptimizer", "LocalMultiGPUOptimizer", "SampleBatch",
     "PolicyEvaluator", "TFMultiGPUSupport", "MultiAgentBatch"]
