+++ python/ray/rllib/optimizers/sync_replay_optimizer.py
@@ -14,7 +14,7 @@ from ray.rllib.utils.filter import RunningStat
 from ray.rllib.utils.timer import TimerStat
 
 
-class LocalSyncReplayOptimizer(PolicyOptimizer):
+class lSyncReplayOptimizer(PolicyOptimizer):
     """Variant of the local sync optimizer that supports replay (for DQN).
 
     This optimizer requires that policy evaluators return an additional
