+++ python/ray/rllib/tuned_examples/mountaincarcontinuous-ddpg.yaml
@@ -61,7 +61,7 @@ mountaincarcontinuous-ddpg:
         # === Parallelism ===
         num_workers: 0
         num_gpus_per_worker: 0
-        optimizer_class: "LocalSyncReplayOptimizer"
+        optimizer_class: "SyncReplayOptimizer"
         optimizer_config: {}
         per_worker_exploration: False
         worker_side_prioritization: False
