+++ python/ray/rllib/tuned_examples/pendulum-ddpg.yaml
@@ -61,7 +61,7 @@ pendulum-ddpg:
         # === Parallelism ===
         num_workers: 0
         num_gpus_per_worker: 0
-        optimizer_class: "LocalSyncReplayOptimizer"
+        optimizer_class: "SyncReplayOptimizer"
         optimizer_config: {}
         per_worker_exploration: False
         worker_side_prioritization: False
